ecmwf-pipeline/requirements.txt
numpy
netcdf4
eccodes
pyyaml
pytest
cdsapi
matplotlib
cartopy
flake8
joblib
scikit-learn
mlflow